Split finance models to separate files
