+++ src/JoinRpg.WebPortal.Models/Money/CreateProjectFeeSettingViewModel.cs
@@ -1,0 +1,7 @@
+namespace JoinRpg.Web.Models
+{
+    public class CreateProjectFeeSettingViewModel : ProjectFeeSettingViewModelBase
+    {
+        public bool PreferentialFeeEnabled { get; set; }
+    }
+}
