+++ src/JoinRpg.WebPortal.Models/Money/EditPaymentTypeViewModel.cs
@@ -1,0 +1,11 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace JoinRpg.Web.Models
+{
+    public class EditPaymentTypeViewModel : PaymentTypeViewModelBase
+    {
+        [Display(Name = "Предлагать по умолчанию")]
+        public bool IsDefault { get; set; }
+        public int PaymentTypeId { get; set; }
+    }
+}
