+++ src/JoinRpg.WebPortal.Models/Money/PaymentTypeViewModelBase.cs
@@ -1,0 +1,11 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace JoinRpg.Web.Models
+{
+    public abstract class PaymentTypeViewModelBase
+    {
+        public int ProjectId { get; set; }
+        [Display(Name = "Название метода оплаты"), Required]
+        public string Name { get; set; }
+    }
+}
